saftbus: move saftbus into saftcommon library
