+++ saftcommon/Connection.cpp
