+++ saftcommon/Connection.h
