+++ saftcommon/Error.cpp
