+++ saftcommon/Error.h
