+++ saftcommon/Interface.cpp
