+++ saftcommon/Interface.h
