+++ saftcommon/Logger.cpp
