+++ saftcommon/MainLoop.cpp
