+++ saftcommon/Makefile.am
@@ -5,7 +5,20 @@ lib_LTLIBRARIES = libsaftcommon.la
 saftcommondir = $(includedir)/saftlib
 saftcommon_HEADERS =         \
 	Time.h             \
-	CommonFunctions.h
+	CommonFunctions.h  \
+	Logger.h           \
+	core.h             \
+	Proxy.h            \
+	Error.h            \
+	Connection.h       \
+	ProxyConnection.h  \
+	Interface.h        \
+	SignalGroup.h      \
+	MainContext.h      \
+	MainLoop.h         \
+	Source.h           \
+	PollFD.h	       \
+	saftbus.h          
 
 leapseconddir = $(datadir)/saftlib
 leapsecond_DATA = leap-seconds.list
@@ -17,6 +30,24 @@ EXTRA_DIST = leap-seconds.list
 libsaftcommon_la_LIBADD = 
 libsaftcommon_la_LDFLAGS = -version-info @SAFTLIB_API@:@SAFTLIB_REVISION@:@SAFTLIB_MINOR@ # @SAFTLIB_SONAME@
 
-libsaftcommon_la_SOURCES = \
-	Time.cpp                \
-	CommonFunctions.cpp
+libsaftcommon_la_SOURCES =     \
+	Time.cpp                   \
+	CommonFunctions.cpp        \
+	Logger.cpp                 \
+	core.cpp                   \
+	Proxy.cpp                  \
+	Error.cpp                  \
+	Connection.cpp             \
+	ProxyConnection.cpp        \
+	Interface.cpp              \
+	SignalGroup.cpp            \
+	saftbus.cpp                \
+	MainContext.cpp            \
+	MainLoop.cpp               \
+	Source.cpp                 \
+	PollFD.cpp     		        
+
+bin_PROGRAMS = saftbus-ctl
+
+saftbus_ctl_SOURCES = saftbus-ctl.cpp
+saftbus_ctl_LDADD   = libsaftcommon.la $(SIGCPP_LIBS)
