+++ saftcommon/SignalGroup.h
