+++ saftcommon/core.cpp
