+++ saftcommon/core.h
