+++ saftcommon/saftbus.cpp
