+++ saftlib.pc.in
@@ -7,5 +7,5 @@ Name: saftlib
 Description: Simplified API for Timing
 Version: @VERSION@
 Requires: sigc++-2.0 >= 0.27.1
-Libs: -L${libdir} -lsaftlib -lsaftbus -lsaftcommon
+Libs: -L${libdir} -lsaftlib -lsaftcommon
 Cflags: -I${includedir}
